requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.26.0
redis==5.0.1
aio-pika==9.4.0
azure-storage-blob==12.19.0
cryptography==42.0.2
elasticsearch==8.12.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
structlog==24.1.0
python-dateutil==2.8.2
